Refactored data to categoriesData
