+++ src/categoriesData.js
